+++ spf/__init__.py
@@ -2,5 +2,5 @@
 from .framework import SanicPluginsFramework
 from .plugin import SanicPlugin
 
-__version__ = '0.4.0.dev20171103'
+__version__ = '0.4.1.dev20171103'
 __all__ = ["SanicPlugin", "SanicPluginsFramework", "__version__"]